requirements.txt
amqp==5.0.9
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
Faker==12.0.1
inflection==0.3.1
iniconfig==1.1.1
isort==5.10.1
kombu==5.2.3
lazy-object-proxy==1.7.1
mixer==7.2.1
more-itertools==8.12.0
packaging==21.3
pluggy==0.13.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==6.2.5
pytest-assert-utils==0.3.0
pytest-common-subject==1.0.5
pytest-django==4.4.0
pytest-fixture-order==0.1.3
pytest-lambda==1.2.5
pytest-pythonpath==0.7.4
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
